Add some v2.0 support to controllers/token
The v2.0 compatible controllers/token function is very simple, and is
stripped for some of the functionality in the v1.0 compatible
controllers/token funtion as checking the claims values and parsing
extra claims. Also, claims are now given as an argument, such that
claims can be configured outside the controllers package.